requirements.txt
argparse
bs4
dnspython
gevent
prettytable
pymongo
requests
scapy
